Create retrain_MIT_Model.py
The primary purpose of making these changes in the threeD_train_final.py file was to retrain the original model on the provided test datasets. Additionally, I wanted to evaluate the classification results of the model, so after saving the checkpoint of the retrain pushUp activity, I loaded this checkpoint and evaluated their results on the pushUp activity dataset. I applied this idea to all the different pose actions and calculated every activity's accuracy separately.
However, a pre-trained model was given but I wanted to retrain it on the available dataset for developing my understanding of the model and the code.