Create keypoint_model_MIT.py
The purpose of making the changes to the threeD_model_final.py was to develop a new model to show the results related to Keypoint instead of Heatmaps. So, instead of following the encoder-decoder structure, I have just kept the encoder part, added a fully connected layer and removed all transpose layers from the model. 
After this, I checked whether this model was giving me the exact results the available model was offering.
I did this just to check the overall performance of the model. Then, I used this model on the pushUp dataset to predict keypoint results.