Create Keypoint_train_test_MIT.py
In this code, I trained and tested the keypoint model on the pushUp activity test dataset and calculated the accuracy. The code focuses on predicting key points only. For this, I had to change the shapes and sizes of the key points.
I applied the keypoint model to every fifteen actions of the human poses to compare their results.
I wanted to compare results like this: available, retrained, and keypoint models.